pyproject.toml
[project]
name = "fyyur-backend-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "babel>=2.17.0",
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-moment>=1.0.6",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.11",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-flask>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.2",
]
